测试微信通知

+++ Jenkinsfile
@@ -28,6 +28,8 @@ pipeline {
         DOCKER_VOLUMES_WORKSPACE_PREFIX = '/data/icql-devops/jenkins/data/workspace'
         JENKINS_WORKSPACE_PREFIX = '/var/jenkins_home/workspace'
         DINGTALK_ROBOT = 'https://oapi.dingtalk.com/robot/send?access_token=8b7536c2584146d4e9d37a8e6b38352adc720990e82f4d09c###isd-400###'
+        WECHAT_ROBOT_ACCESS_TOKEN = 'https://qyapi.weixin.qq.com/cgi-bin/gettoken?corpid=###isd-500###&corpsecret=###isd-510###'
+        WECHAT_ROBOT = 'https://qyapi.weixin.qq.com/cgi-bin/message/send?access_token='
     }
 
     stages {
@@ -99,7 +101,7 @@ pipeline {
                                     credentialsId: 'icql-secret-dictionary',
                                     usernameVariable: 'ICQL_SECRET_KEY',
                                     passwordVariable: 'ICQL_SECRET_VALUE')]) {
-                                def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll('{', '[').replaceAll('}', ']'))
+                                def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll("\\{", "[").replaceAll("\\}", "]"))
                                 secretWordsMap.each { key, value ->
                                     sh "sed -i \"s/${value}/******/g\" `grep \"${value}\" -rl ${JENKINS_WORKSPACE_PREFIX}/00_ICQL/deploy-static/00_home/public` || true"
                                 }
@@ -137,7 +139,7 @@ pipeline {
                                         credentialsId: 'icql-secret-dictionary',
                                         usernameVariable: 'ICQL_SECRET_KEY',
                                         passwordVariable: 'ICQL_SECRET_VALUE')]) {
-                                    def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll('{', '[').replaceAll('}', ']'))
+                                    def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll("\\{", "[").replaceAll("\\}", "]"))
                                     secretWordsMap.each { key, value ->
                                         sh "sed -i \"s/${key}/${value}/g\" `grep \"${key}\" -rl ./conf/${namespace}` || true"
                                     }
@@ -214,24 +216,52 @@ pipeline {
 }
 
 def sendMessage(result) {
-    //组装钉钉通知内容
-    def latestCommits = ''
-    if (fileExists("${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git/${GIT_REPONAME}_LATEST_COMMITS")) {
-        latestCommits = readFile("${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git/${GIT_REPONAME}_LATEST_COMMITS").trim()
-        sh "rm -rf ${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git"
-    }
-    def message = "{\"msgtype\":\"markdown\",\"markdown\":{\"title\":\"DS 通知\",\"text\":\"" +
-            "### [${JOB_NAME}/${BUILD_NUMBER}](${BUILD_URL}/console) ${result}\\n" +
-            "#### 最近更新的内容：\\n" +
-            "${latestCommits}\\n" +
-            "\"}}"
-    //发送钉钉通知
     withCredentials([usernamePassword(
             credentialsId: 'icql-secret-dictionary',
             usernameVariable: 'ICQL_SECRET_KEY',
             passwordVariable: 'ICQL_SECRET_VALUE')]) {
-        def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll('{', '[').replaceAll('}', ']'))
+        def secretWordsMap = evaluate(ICQL_SECRET_VALUE.replaceAll("\\{", "[").replaceAll("\\}", "]"))
+
+        //组装钉钉通知内容
+        def latestCommits = ''
+        if (result == '成功' && fileExists("${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git/${GIT_REPONAME}_LATEST_COMMITS")) {
+            latestCommits = readFile("${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git/${GIT_REPONAME}_LATEST_COMMITS").trim()
+            sh "rm -rf ${JENKINS_WORKSPACE_PREFIX}/00_ICQL/tmp-git"
+        }
+        def dingMessage = "{\"msgtype\":\"markdown\",\"markdown\":{\"title\":\"DS 通知\",\"text\":\"" +
+                "### [${JOB_NAME}/${BUILD_NUMBER}](${BUILD_URL}/console) ${result}\\n" +
+                "#### 最近更新的内容：\\n" +
+                "${latestCommits}\\n" +
+                "\"}}"
+        //发送钉钉通知
         def dingRobot = DINGTALK_ROBOT.replaceAll("###isd-400###", secretWordsMap["###isd-400###"])
-        sh "curl ${dingRobot} -H 'Content-Type:application/json' -X POST --data '${message}'"
+        sh "curl ${dingRobot} -H 'Content-Type:application/json' -X POST --data '${dingMessage}'"
+
+        //组装微信通知内容
+        def wechatMessage = "{\n" +
+                "    \"toparty\": \"1\",\n" +
+                "    \"agentid\": ${secretWordsMap["###isd-511###"]},\n" +
+                "    \"msgtype\": \"news\",\n" +
+                "    \"news\": {\n" +
+                "        \"articles\": [\n" +
+                "            {\n" +
+                "                \"title\": \"Deploy-Static通知\",\n" +
+                "                \"description\": \"[${JOB_NAME}/${BUILD_NUMBER}] ${result}\",\n" +
+                "                \"url\": \"${BUILD_URL}\",\n" +
+                "                \"picurl\": \"https://file.icql.work/30_picture/1500_jenkins.jpg\"\n" +
+                "            }\n" +
+                "        ]\n" +
+                "    },\n" +
+                "    \"safe\": 0,\n" +
+                "    \"enable_id_trans\": 0,\n" +
+                "    \"enable_duplicate_check\": 0,\n" +
+                "    \"duplicate_check_interval\": 1800\n" +
+                "}"
+        //获取微信应用access_token
+        sh "curl -s -- ${WECHAT_ROBOT_ACCESS_TOKEN} > tmp-WECHAT_ROBOT_ACCESS_TOKEN"
+        def wechatRobotAccessToken = readFile('tmp-WECHAT_ROBOT_ACCESS_TOKEN').trim()
+        sh 'rm -rf tmp-*'
+        //发送微信消息通知
+        sh "curl ${WECHAT_ROBOT}${wechatRobotAccessToken} -H 'Content-Type:application/json' -X POST --data '${wechatMessage}'"
     }
 }
